# api/openapi/stackit.yaml
openapi: 3.1.0
info:
  title: Stackit API
  version: 1.0.0
  description: API used by the in-repo Stackit web frontend.
servers:
  - url: /
paths:
...
  /api/v1/stacks/{rootBranch}:
    get:
      summary: Get detailed stack by root branch.
      operationId: getStack
      parameters:
        - in: path
          name: rootBranch
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stack details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StackDetail"
        "404":
          description: Stack not found.
...
  /api/v1/branches/{branchName}:
    get:
      summary: Get one tracked branch.
      operationId: getBranch
      parameters:
        - in: path
          name: branchName
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Branch details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BranchResponse"
        "404":
          description: Branch not found.
...
components:
  schemas:
...
    StackSummary:
      type: object
      required: [rootBranch, title, status, branchCount, prCount, isCurrent]
      properties:
        rootBranch:
          type: string
        title:
          type: string
        status:
          type: string
        scope:
          type: string
        branchCount:
          type: integer
        prCount:
          type: integer
        isCurrent:
          type: boolean
        description:
          type: string
        owner:
          type: string
    StackDetail:
      allOf:
        - $ref: "#/components/schemas/StackSummary"
        - type: object
          required: [branches]
          properties:
            branches:
              type: array
              items:
                $ref: "#/components/schemas/BranchResponse"
    BranchResponse:
      type: object
      required: [name, depth, isCurrent, needsRestack, isLocked, isFrozen, revision]
      properties:
        name:
          type: string
        parent:
          type: string
        children:
          type: array
          items:
            type: string
        depth:
          type: integer
        isCurrent:
          type: boolean
        needsRestack:
          type: boolean
        isLocked:
          type: boolean
        lockReason:
          type: string
        isFrozen:
          type: boolean
        scope:
          type: string
        revision:
          type: string
        commitDate:
          type: string
        commitAuthor:
          type: string
        commitCount:
          type: integer
        linesAdded:
          type: integer
        linesDeleted:
          type: integer
        commits:
          type: array
          items:
            $ref: "#/components/schemas/CommitResponse"
        pr:
          $ref: "#/components/schemas/PRResponse"
        ci:
          $ref: "#/components/schemas/CIResponse"
    CommitResponse:
      type: object
      required: [sha, message]
      properties:
        sha:
          type: string
        message:
          type: string
    PRResponse:
      type: object
      required: [number, title, state, url, isDraft, base]
      properties:
        number:
          type: integer
        title:
          type: string
        state:
          type: string
        url:
          type: string
        isDraft:
          type: boolean
        base:
          type: string
    CIResponse:
      type: object
      required: [status, reviewDecision]
      properties:
        status:
          type: string
        reviewDecision:
          type: string
        checks:
          type: array
          items:
            $ref: "#/components/schemas/CheckDetailResponse"
    CheckDetailResponse:
      type: object
      required: [name, status, conclusion]
      properties:
        name:
          type: string
        status:
          type: string
        conclusion:
          type: string